GUI for subtitle sync in the "Track synchronization" under OSX : *NOT* done
As this file depends on Interface Builder 3, I could not update it
in order to add the subtitle sync feature which I added in the
Qt "Track Synchronization window".

I prefer not to mess around it, as I am not specialist enough
with interface builder...

Instead, I added comments that explain which buttons should be added and
what their behaviour should be.